+++ src/layouts/builder.vue
@@ -1,0 +1,81 @@
+<template>
+  <q-layout> <!-- Be sure to play with the Layout demo on docs -->
+
+    <!-- (Optional) The Header -->
+    <q-layout-header>
+      <q-toolbar>
+        <q-btn
+          flat
+          round
+          dense
+          icon="menu"
+          @click="leftDrawer = !leftDrawer"
+        />
+        <q-toolbar-title>
+          Header
+          <span slot="subtitle">Subtile</span>
+        </q-toolbar-title>
+      </q-toolbar>
+    </q-layout-header>
+
+    <!-- (Optional) The Footer -->
+    <q-layout-footer>
+      <q-toolbar>
+        <q-btn
+          flat
+          round
+          dense
+          icon="menu"
+          @click="leftDrawer = !leftDrawer"
+        />
+        <q-toolbar-title>
+          Footer
+          <span slot="subtitle">Subtile</span>
+        </q-toolbar-title>
+      </q-toolbar>
+    </q-layout-footer>
+
+    <!-- (Optional) A Drawer; you can add one more with side="right" or change this one's side -->
+    <q-layout-drawer
+      side="left"
+      v-model="leftDrawer"
+      content-class="bg-grey-9"
+    >
+      <!-- QScrollArea is optional -->
+      <q-scroll-area class="fit q-pa-sm">
+        <q-item link @click.native="logout">
+          <q-item-side icon="fa-sign-out fab" />
+          <q-item-main label="Log Out" />
+        </q-item>
+      </q-scroll-area>
+    </q-layout-drawer>
+
+    <q-page-container>
+      <!-- This is where pages get injected -->
+      <router-view />
+    </q-page-container>
+
+  </q-layout>
+</template>
+
+<script>
+export default {
+  // name: 'LayoutName',
+  data () {
+    return {
+      leftDrawer: true
+    }
+  },
+  methods: {
+    logout () {
+      console.log('signing out')
+      this.firebase.auth.signOut().then(() => {
+        this.$router.replace('login')
+      })
+    }
+  }
+}
+</script>
+
+<style>
+</style>