+++ src/router/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
+import firebase from 'firebase'
 
 import routes from './routes'
 
@@ -19,6 +20,21 @@ const Router = new VueRouter({
   base: process.env.VUE_ROUTER_BASE,
   scrollBehavior: () => ({ y: 0 }),
   routes
+})
+
+Router.beforeEach((to, from, next) => {
+  console.log('router before each')
+  if (to.matched.some(record => record.meta.requiresAuth)) {
+    firebase.auth().onAuthStateChanged(function (user) {
+      if (!user) {
+        console.log('no user')
+        next('login')
+      } else {
+        console.log(user)
+        next()
+      }
+    })
+  } else next()
 })
 
 export default Router