+++ src/router/routes.js
@@ -1,10 +1,23 @@
 
 export default [
+  { path: '/', redirect: '/dashboard' },
+
   {
     path: '/',
+    component: () => import('layouts/builder'),
+    children: [
+      { path: 'dashboard', name: 'dashboard', component: () => import('pages/Dashboard') }
+    ],
+    meta: {
+      requiresAuth: true
+    }
+  },
+
+  {
+    path: '/login',
     component: () => import('layouts/default'),
     children: [
-      { path: '', component: () => import('pages/index') }
+      { path: '', component: () => import('pages/Login') }
     ]
   },