+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes for this project will be documented in this file.
 
+## 0.0.2 - 2018-03-19
+### Added
+- Login page
+- Dashboard page
+- Logout button on the builder sidebar
+- firebase plugin - includes vuefire and vue-firestore
+- vuelidate plugin
+- REAL Curriculum logo
+
+### Changed
+- added Firebase authentication to router
+- set dark theme
+
 ## 0.0.1 - 2018-03-19
 ### Added
 - Initial commit